refactor: modularize dashboard utilities and components
