+++ src/pages/Dashboard/components/LayerCard.module.css
@@ -1,0 +1,61 @@
+.card {
+    background: #fff;
+    border: 0.0625rem solid #e5e7eb;
+    border-radius: 0.75rem;
+    padding: 0.875rem;
+    margin-bottom: 1rem;
+}
+
+.layer {
+    cursor: pointer;
+    transition: transform .15s ease;
+    position: relative;
+    border-left: 0.125rem solid #90c398;
+    padding-top: 0rem;
+}
+
+.layer:hover {
+    transform: translateY(-0.125rem);
+}
+
+.headerRow { display:flex; justify-content: space-between; align-items:center; cursor:pointer; }
+
+.dot {
+    width: 0.5rem;
+    height: 0.5rem;
+    border-radius: 6.1875rem;
+    display: inline-block;
+}
+
+.ok { background: #10b981; }
+.warn { background: #f59e0b; }
+.down { background: #ef4444; }
+
+.stats {
+    display: flex;
+    gap: 0.75rem;
+    flex-wrap: wrap;
+    margin-top: 0.5rem;
+}
+
+.details {
+    margin-top: 0.5rem;
+    font-size: .9rem;
+    color: #374151;
+    border-top: 0.0625rem dashed #e5e7eb;
+    padding-top: 0.5rem;
+}
+
+.devCards {
+    display: grid;
+    grid-template-columns: 1fr;
+    gap: 1rem;
+    justify-items: stretch;
+    padding: 0 0.5rem;
+    box-sizing: border-box;
+}
+
+.muted {
+    color: #6b7280;
+    font-size: .9rem;
+}
