+++ src/pages/Dashboard/components/Stat.module.css
@@ -1,0 +1,26 @@
+.stat {
+    background: #f3f4f6;
+    border-radius: 0.5rem;
+    padding: 0.375rem 0.625rem;
+    font-size: .92rem;
+}
+
+.statOk {
+    background: linear-gradient(180deg, #8deaa3 0%, #a7e6b2 100%);
+    color: #000;
+}
+
+.statWarn {
+    background: #fff3cd;
+    color: #000;
+}
+
+.statDanger {
+    background: #f8d7da;
+    color: #000;
+}
+
+.muted {
+    color: #6b7280;
+    font-size: .9rem;
+}
